create static global variable to handle exposed public keys related to a certain trusted registry
